social login

+++ ducttapp/api/user/user.py
@@ -1,9 +1,9 @@
 from flask_restplus import Resource, fields
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from flask import request
 from . import ns
 
-from ducttapp import models, services
+from ducttapp import models, services, repositories, extensions
 from ducttapp.helpers.decorators import user_management_required
 
 role_model = ns.model(
@@ -39,12 +39,25 @@ class User(Resource):
     @user_management_required
     @ns.marshal_with(_update_user_res)
     def put(self, user_id):
+        current_user = repositories.user.find_one_by_email_or_username_in_user_ignore_case(
+            username=get_jwt_identity()
+        )
+        if current_user and current_user.id == user_id:
+            raise extensions.exceptions.ForbiddenException(
+                message="You can't edit myself"
+            )
         result = services.admin.edit_user(user_id, request.form)
         return result
 
     @jwt_required
     @user_management_required
-    @ns.doc('delete user with id')
     def delete(self, user_id):
+        current_user = repositories.user.find_one_by_email_or_username_in_user_ignore_case(
+            username=get_jwt_identity()
+        )
+        if current_user and current_user.id == user_id:
+            raise extensions.exceptions.ForbiddenException(
+                message="You can't delete myself"
+            )
         message = services.admin.delete_user(user_id)
         return message
